requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
decorator==4.4.2
future==0.18.2
idna==2.10
lxml==4.6.1
pycparser==2.20
python-telegram-bot==12.7
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.11
